chore: remove ability to use Django models in streamlit app
this setup was causing an error when deploying to AWS since the Django
pems app resides in a different container. when running a local
debugger though, this error was not present since the local dev
container does have the pems app. since at this moment we are not
using Django models in the streamlit app, we are removing this ability.