chore(commands): remove redundant database setup commands
since ensure_db implements the same functionality, the .sh commands
are cleaned up to avoid redundancy.
in the pytest workflow, running migrations is not needed since
tests will migrate the database, but we still need to collect static
files and generate the static files manifest so that the
workflow does not fail.