fix(infra): get environment name from vars
although some references suggested that `github.environment` could
be used to obtain the environment's name, testing the workflow showed
that `github.environment` is actually not available in the context.
instead, we set ENVIRONMENT_NAME for each environment and use
`vars.ENVIRONMENT_NAME` to get the name of the deployment environment
for the workflow.