chore(5-min-viz): clear out stale district/station info
if a station is selected and then de-selected, some information
associated with the station is not cleared, i.e., Absolute Post Mile
and Lanes stay greyed out, Streamlit's sign of "stale data". it turns
out that the st.empty() method sometimes needs some time to take effect
on the front-end. adding a small delay ensures that the DOM is ready
for st.empty().